requirements.txt
python-telegram-bot
glob3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib